- added resize event

+++ 2021Fall/src/Engine/Graphic/GraphicPipeline.cpp
@@ -109,7 +109,7 @@ void GraphicPipeline::init(VkRenderPass renderpass, VkDescriptorSetLayout descri
 
 	VkGraphicsPipelineCreateInfo pipelineInfo{};
 	pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-	pipelineInfo.stageCount = static_cast<size_t>(shaderStages.size());
+	pipelineInfo.stageCount = static_cast<uint32_t>(shaderStages.size());
 	pipelineInfo.pStages = shaderStages.data();
 	pipelineInfo.pVertexInputState = &inputstate;
 	pipelineInfo.pInputAssemblyState = &inputAssembly;
